+++ R/scale.R
@@ -227,6 +227,26 @@ scale_level <- function(data) {
 }
 
 
+#' Scale intensities in MSn spectra to the highest value within each spectrum
+#'
+#' @description
+#' Scale the intensity of each peak in an MSn spectrum to that of the highest peak. MSn spectra are required to use this function.
+#' See \code{\link[metamorphr]{read_mgf}}.
+#'
+#' \strong{Important Note}
+#'
+#' Please note that existing MSn spectra in `data` will be overwritten.
+#'
+#'
+#' @param data A tidy tibble created by \code{\link[metamorphr]{read_featuretable}}.
+#' @param scale_to A `numeric` that specifies to which number the highest signal in each spectrum will be scaled.
+#'
+#' @returns A tibble with scaled MSn spectra.
+#' @export
+#'
+#' @examples
+#' toy_mgf %>%
+#'   scale_msn()
 scale_msn <- function(data, scale_to = 100) {
   col_order <- names(data)