+++ tests/testthat/test-calc_neutral_loss.R
@@ -1,6 +1,6 @@
 test_that("row & column order stays unchanged", {
   mgf_tibble <- read_mgf(test_path("data", "test_filter_msn_nl.mgf"))
-  mgf_tibble_nl <- calc_neutral_loss(mgf_tibble, m_z_col = PEPMASS)
+  mgf_tibble_nl <- msn_calc_nl(mgf_tibble, m_z_col = PEPMASS)
 
   expect_equal(dplyr::select(mgf_tibble_nl, -"Neutral_Loss"), mgf_tibble)
 })
@@ -29,7 +29,7 @@ test_that("neutral_losses are correct", {
     211.2222,13)
 
   mgf_tibble <- read_mgf(test_path("data", "test_filter_msn_nl.mgf"))
-  mgf_tibble_nl <- calc_neutral_loss(mgf_tibble, m_z_col = PEPMASS) %>%
+  mgf_tibble_nl <- msn_calc_nl(mgf_tibble, m_z_col = PEPMASS) %>%
     dplyr::pull("Neutral_Loss")
 
   expect_equal(mgf_tibble_nl[[1]], nl1)
@@ -40,7 +40,7 @@ test_that("neutral_losses are correct", {
 test_that("returns NULL if no MSn spectrum is available", {
   mgf_tibble <- read_mgf(test_path("data", "test_filter_msn_nl.mgf"))
   mgf_tibble$MSn[1] <- list(NULL)
-  mgf_tibble_nl <- calc_neutral_loss(mgf_tibble, m_z_col = PEPMASS)
+  mgf_tibble_nl <- msn_calc_nl(mgf_tibble, m_z_col = PEPMASS)
 
   expect_null(mgf_tibble$MSn[[1]])
   expect_null(mgf_tibble_nl$Neutral_Loss[[1]])
@@ -49,7 +49,13 @@ test_that("returns NULL if no MSn spectrum is available", {
 test_that("returns NULL if Neutral_Loss tibble is empty", {
   mgf_tibble <- read_mgf(test_path("data", "test_filter_msn_nl.mgf"))
   mgf_tibble$PEPMASS[1] <- 1
-  mgf_tibble_nl <- calc_neutral_loss(mgf_tibble, m_z_col = PEPMASS)
+  mgf_tibble_nl <- msn_calc_nl(mgf_tibble, m_z_col = PEPMASS)
 
   expect_null(mgf_tibble_nl$Neutral_Loss[[1]])
+})
+
+test_that("deprecation warning when calling old function", {
+  mgf_tibble <- read_mgf(test_path("data", "test_filter_msn_nl.mgf"))
+
+  expect_warning(calc_neutral_loss(mgf_tibble, m_z_col = PEPMASS))
 })