+++ NEWS.md
@@ -1,3 +1,5 @@
+# metamorphr (development version)
+
 # metamorphr 0.2.0
 
 ## New features